// Portfolio/src/app/interfaces/itokenpayload.ts
export interface ITokenPayload {
  aud?: string | string[] | undefined;  //(audience) claim
  exp?: number | undefined;             //(Expiration time) claim
  iat?: number | undefined;             //(IssuedAt time) claim
  iss?: string | undefined;             //(issuer) claim
  jti?: string | undefined;             //() claim
  nbf?: number | undefined;             //(Not-Before time) claim
  sub?: string | undefined;             //(Subject) claim

  id?: number | undefined;
  rule?: string | undefined;
  firstname?: string | undefined;
  lastname?: string | undefined;
}